Move trainer functions to trainer.c (backport)
