+++ radio/src/targets/flysky/board.cpp
@@ -308,64 +308,11 @@ void boardOff()
   lcdOff();
   SysTick->CTRL = 0; // turn off systick
   pwrOff();
-}
 
-uint8_t currentTrainerMode = 0xff;
+  // disable interrupts
+  __disable_irq();
 
-void checkTrainerSettings()
-{
-  uint8_t requiredTrainerMode = g_model.trainerMode;
-  if (requiredTrainerMode != currentTrainerMode)
-  {
-    switch (currentTrainerMode)
-    {
-    case TRAINER_MODE_MASTER_TRAINER_JACK:
-      //stop_trainer_capture();
-      break;
-    /*
-    case TRAINER_MODE_SLAVE:
-      stop_trainer_ppm();
-      break;
-    case TRAINER_MODE_MASTER_CPPM_EXTERNAL_MODULE:
-      stop_cppm_on_heartbeat_capture() ;
-      break;
-    case TRAINER_MODE_MASTER_SBUS_EXTERNAL_MODULE:
-      stop_sbus_on_heartbeat_capture() ;
-      break;*/
-#if defined(TRAINER_BATTERY_COMPARTMENT)
-    case TRAINER_MODE_MASTER_BATTERY_COMPARTMENT:
-      auxSerialStop();
-      break;
-#endif
-    }
-
-    currentTrainerMode = requiredTrainerMode;
-    switch (requiredTrainerMode) {
-    /*
-    case TRAINER_MODE_SLAVE:
-      init_trainer_ppm();
-      break;
-    case TRAINER_MODE_MASTER_CPPM_EXTERNAL_MODULE:
-        init_cppm_on_heartbeat_capture();
-        break;
-    case TRAINER_MODE_MASTER_SBUS_EXTERNAL_MODULE:
-        init_sbus_on_heartbeat_capture();
-        break;
-    */
-#if defined(TRAINER_BATTERY_COMPARTMENT)
-    case TRAINER_MODE_MASTER_BATTERY_COMPARTMENT:
-      if (g_eeGeneral.auxSerialMode == UART_MODE_SBUS_TRAINER) {
-        auxSerialSbusInit();
-        break;
-      }
-      // no break
-#endif
-    default:
-      // master is default
-      init_trainer_capture();
-      break;
-    }
-  }
+  // this function must not return!
 }
 
 uint16_t getBatteryVoltage()
