--- radio/src/trainer_input.cpp
@@ -1,57 +1,0 @@
-/*
- * Copyright (C) OpenTX
- *
- * Based on code named
- *   th9x - http://code.google.com/p/th9x 
- *   er9x - http://code.google.com/p/er9x
- *   gruvin9x - http://code.google.com/p/gruvin9x
- *
- * License GPLv2: http://www.gnu.org/licenses/gpl-2.0.html
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2 as
- * published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- */
-
-#include "trainer_input.h"
-
-int16_t ppmInput[MAX_TRAINER_CHANNELS];
-uint8_t ppmInputValidityTimer;
-
-
-#if defined(STM32F0)
-void checkTrainerSignalWarning()
-{
-	
-}
-#else
-#include "audio.h"
-
-void checkTrainerSignalWarning()
-{
-  enum PpmInValidState_t {
-    PPM_IN_IS_NOT_USED=0,
-    PPM_IN_IS_VALID,
-    PPM_IN_INVALID
-  };
-
-  static uint8_t ppmInputValidState = PPM_IN_IS_NOT_USED;
-
-  if (ppmInputValidityTimer && (ppmInputValidState == PPM_IN_IS_NOT_USED)) {
-    ppmInputValidState = PPM_IN_IS_VALID;
-  }
-  else if (!ppmInputValidityTimer && (ppmInputValidState == PPM_IN_IS_VALID)) {
-    ppmInputValidState = PPM_IN_INVALID;
-    AUDIO_TRAINER_LOST();
-  }
-  else if (ppmInputValidityTimer && (ppmInputValidState == PPM_IN_INVALID)) {
-    ppmInputValidState = PPM_IN_IS_VALID;
-    AUDIO_TRAINER_BACK();
-  }
-}
-#endif
